blood bank/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-Caching==2.1.0
Flask-Security==5.3.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Admin==1.6.1
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.5
Pillow==10.0.1
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.0
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
gevent==23.9.1
eventlet==0.33.3
Jinja2==3.1.3
itsdangerous==2.1.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 
email-validator==2.1.0.post1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
isort==5.13.2 